moved Exhibition Resource parser
